New stylesheet, transcription corrections
I have re-designed the stylesheet used ti generate detailed general orpus data. It is now faster (approx 4 hours, as opposed to 36!) and the data is split up for better searching. And the context is rationalised, meaning that more or less is retrieved, depending on the position of the word.

In the process, I also found some typos etc. in the transcription files.