Bug fix
The unique number assigned to a comment row was ceasing to be unique when the html file was saved and then reloaded. A "session number" has now been added to the <html> element, which will update each time the HTML document is saved or refreshed. This will then be used as a component in the id for each comment.