Clickthrough Function
In semi-dip and dip editions, in default mode, clicking on a word will now simply take you straight to the dictionary entry pointed to by @lemmaRef. When Collect Slips or Headword Search is enabled, clicking will add a slip or slips to a new table. When a mode is disabled, direct clickthrough to a dictionary resumes.